Split media library server helpers
